More re-org, move everything to separate classes
