+++ src/storage/CMakeLists.txt
@@ -1,4 +1,12 @@
-add_library(quack_storage OBJECT quack_catalog.cpp quack_insert.cpp)
+add_library(
+  quack_storage OBJECT
+  quack_catalog.cpp
+  quack_insert.cpp
+  quack_table.cpp
+  quack_schema.cpp
+  quack_transaction.cpp
+  quack_transaction.cpp
+  quack_transaction_manager.cpp)
 set(EXTENSION_SOURCES
     ${EXTENSION_SOURCES} $<TARGET_OBJECTS:quack_storage>
     PARENT_SCOPE)
